resume_builder_agent.py
from openai import AsyncOpenAI
from typing import Dict

class ResumeBuilderAgent:
    def __init__(self, openai_api_key: str):
        self.client = AsyncOpenAI(api_key=openai_api_key)

    async def create_resume(self, user_data: dict) -> str:
        # Implementation using OpenAI
        return "Generated Resume"

    async def tailor_resume(self, resume: str, job_description: str) -> str:
        # Implementation using OpenAI
        return "Tailored Resume"

    async def analyze_resume(self, resume: str) -> Dict:
        # Implementation using OpenAI
        return {
            "strengths": ["Strong technical skills", "Clear experience"],
            "improvements": ["Add more quantifiable achievements"]
        } 